Fixed testing and dependencies (#9)
# Description

Fixed so tests work as they should and added dependencies.
<!--
Please include a summary of the change and which issue(s) are fixed.
Please also include relevant motivation and context. List any
dependencies that are required for this change.
-->

<!-- If this pull request fixes one or more issues, uncomment the line
below
     and replace 123 with the relevant issue number(s), for example:
     "Fixes #123" or "Fixes #123, Fixes #456".
Fixes #123
-->

## Type of change

<!--
Please delete options that are not relevant.
-->

- [X] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

# How Has This Been Tested?

<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration.
-->

- [X] Ran all defined tests.

**Test Configuration**:
* OS: Ubuntu 24.04

# Checklist

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules

---------

+++ lib/container/include/ufo/container/structure_of_arrays/element.hpp
@@ -324,7 +324,7 @@ class SoAElement
 	[[nodiscard]] std::size_t position() const
 	{
 		assert(is_reference_ && nullptr != reference_);
-		return std::distance(reference_->cbegin(), this);
+		return static_cast<std::size_t>(std::distance(reference_->cbegin(), this));
 	}
 
  private:
